+++ background-geolocation.ios.d.ts
@@ -20,7 +20,8 @@ export declare class BackgroundGeolocation extends AbstractBackgroundGeolocation
     static watchPosition(success: Function, failure?: Function, options?: Object): void;
     static stopWatchPosition(success?: Function, failure?: Function): void;
     static getOdometer(success: Function, failure?: Function): void;
-    static resetOdometer(success?: Function): void;
+    static setOdometer(value: number, success?: Function, failure?: Function): void;
+    static resetOdometer(success?: Function, failure?: Function): void;
     /**
     * HTTP & Persistence Methods
     */